Add countdown

+++ db/migrate/20230816095938_add_exchange_rate_column_to_appointments.rb
@@ -1,5 +1,5 @@
 class AddExchangeRateColumnToAppointments < ActiveRecord::Migration[7.0]
   def change
-    add_column :appointments, :exchange_rate, :integer
+    add_column :appointments, :exchange_rate, :float
   end
 end
